Add __isset magic method
When doing custom pagination searches, it is handy to use isset() functions to see if ranges are set (members;range=0-499 vs members;range=0-*).  This simple __isset method should allow that to work.

I had to get the attribute first.  Then do the isset.  I think because it would create a loop otherwise.